requirements.txt
streamlit
python-dotenv
langchain
langchain_google_genai
langgraph
langchain-community
duckduckgo-search
wikipedia